requirements.txt
# Core
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0

# Feishu
feishu>=0.1.0
websocket-client>=1.0.0

# Web
flask>=3.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
eventlet>=0.35.0

# Utils
rich>=13.0.0
tqdm>=4.0.0
